move get media calls to client
